% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_birdnet_py.R
\name{call_birdnet.py}
\alias{call_birdnet.py}
\title{Return a Python call to analyse Audio files with BirdNET}
\usage{
call_birdnet.py(
  input,
  output,
  overlap = 2,
  threads = NULL,
  batch_size = NULL,
  sensitivity = 1,
  slist,
  locale = "en",
  lat = -1,
  lon = -1,
  week = -1,
  min_conf = 0.1,
  rtype = "table",
  fmin = 0,
  fmax = 15000,
  audio_speed = 1,
  combine_results = NULL
)
}
\arguments{
\item{input}{Character. File path to audio files}

\item{output}{Character. File path to target directory.}

\item{overlap}{Numeric. See description of BirdNET software for valid values.}

\item{threads}{Integer. See description of BirdNET software for valid values.}

\item{batch_size}{Integer. See description of BirdNET software for valid values.}

\item{sensitivity}{Numeric. See description of BirdNET software for valid values.}

\item{slist}{Character. Specify path to species list.}

\item{locale}{Character. See description of BirdNET software for valid values.}

\item{lat}{Numeric. See description of BirdNET software for valid values.}

\item{lon}{Numeric. See description of BirdNET software for valid values.}

\item{week}{Integer. See description of BirdNET software for valid values.}

\item{min_conf}{Numeric. See description of BirdNET software for valid values.}

\item{rtype}{Character. See description of BirdNET software for valid values.}

\item{fmin}{Integer. See description of BirdNET software for valid values.}

\item{fmax}{Integer. See description of BirdNET software for valid values.}

\item{audio_speed}{Numeric. See description of BirdNET software for valid values.}

\item{combine_results}{Character. See description of BirdNET software for valid values.}
}
\description{
Return a Python call to analyse Audio files with BirdNET
}
