% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clipcat_birdnet_selection.R
\name{clipcat_birdnet_selection}
\alias{clipcat_birdnet_selection}
\title{Clip and concatenate WAV-files and selection tables}
\usage{
clipcat_birdnet_selection(
  df,
  output_dir,
  sec_before = 5,
  sec_after = 12,
  birdnet_clip_length = 3,
  dur_silence = 0.5
)
}
\arguments{
\item{df}{Character. Dataframe of a birdnet selection table inclduing the follwing variables: Selection, View, Channel, `Begin Path`, `Begin Time (s)`, `End Time (s)`, `Low Freq (Hz)`, `High Freq (Hz)`, Confidence, `File Offset (s)`, rowid, `Common Name`.}

\item{output_dir}{Character. Directory for the output.}

\item{sec_before}{Numeric. Time (in seconds) that should be added to the clipped Audio before the actual snippet of interest.}

\item{sec_after}{Numeric. Time (in seconds) that should be added to the clipped Audio after the actual snippet of interest.}

\item{birdnet_clip_length}{Numeric. Clip length of the snippet of interest in seconds.}

\item{dur_silence}{Numeric. Time (in seconds) with silence between each clip.}
}
\description{
Clip and concatenate WAV-files and selection tables
}
