% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_dataframes.R
\name{combine_dataframes}
\alias{combine_dataframes}
\title{Combine multiple similar dataframes}
\usage{
combine_dataframes(file_paths)
}
\arguments{
\item{file_paths}{Character vector. Specify a vector of filepaths to multiple dataframes. Intended usage is to combine BirdNET-results by using the column 'filepath.birdnet.selection.table' from the log-file.}
}
\description{
A wrapper function that reads and combines dataframes with the data.table-package.
}
