% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config2df.R
\name{config2df}
\alias{config2df}
\title{Tabularize Audiomoth configuration file}
\usage{
config2df(filepath.config, filepath.export = NULL)
}
\arguments{
\item{filepath.config}{Character. Specify filepath to the directory containing all files copied from the microSD-card of the AudioMoth.}

\item{filepath.export}{Character. Optionally specify filepath to export results as csv.}
}
\description{
Tabularize Audiomoth configuration file
}
