% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dirs.R
\name{create_dirs}
\alias{create_dirs}
\title{Create directories for audio files}
\usage{
create_dirs(filepath.target, YMDstart, YMDend, content)
}
\arguments{
\item{filepath.target}{Character. Specify the full file path to the directory for the sample. The last subdirectory should designate a unique identifier for the location of the device that recorded the audio files.}

\item{YMDstart}{Character. Specify the date of the first recoring in the format YYYYmmdd.}

\item{YMDend}{Character. Specify the date of the last recoring in the format YYYYmmdd.}

\item{content}{Character. User-defined specification for the target directory.}
}
\description{
Create directories for audio files
}
