% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_metadata.R
\name{get_metadata}
\alias{get_metadata}
\title{Wrapper for reading and writing GUANO metadata}
\usage{
get_metadata(filepath.source, filepath.target)
}
\arguments{
\item{filepath.source}{Character. Specify directory which contains GUANO files.}

\item{filepath.target}{Character.}
}
\description{
Wrapper for reading and writing GUANO metadata
}
