% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_recording_duration.R
\name{get_recording_duration}
\alias{get_recording_duration}
\title{get first and last recording date}
\usage{
get_recording_duration(filepath.source)
}
\arguments{
\item{filepath.source}{Character. Specify the full file path to the directory where the AudioMoth files are stored (e.g. the full path to the microSD card). The directory should contain daily folders with names of dates in the format YYYYmmdd.}
}
\description{
get first and last recording date
}
