% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_metadata.R
\name{visualize_metadata}
\alias{visualize_metadata}
\title{Plot predefined figures of metadata}
\usage{
visualize_metadata(filepath.metadata)
}
\arguments{
\item{filepath.metadata}{Character. Specify file path to csv containing GUANO metadata.}
}
\description{
Plot predefined figures of metadata
}
