% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_log.R
\name{write_log}
\alias{write_log}
\title{Write a log file of the specified parameters}
\usage{
write_log(filepath.logfile = NULL, filepath.source, filepath.target)
}
\arguments{
\item{filepath.logfile}{Character. Specify file path to a log file. A new file will be written, if missing.}

\item{filepath.source}{Character. Specify the variable filepath.source.}

\item{filepath.target}{Character. Specify the variable filepath.target.}
}
\description{
Write a log file of the specified parameters
}
