% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_logfile.R
\name{update_logfile}
\alias{update_logfile}
\title{Add current timestamp to logfile}
\usage{
update_logfile(UID, filepath.logfile)
}
\arguments{
\item{UID}{Integer. Unique Identifier.}

\item{filepath.logfile}{Character. Filepath to logfile.csv}
}
\description{
Add current timestamp to logfile
}
