% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_metadata.R
\name{visualize_metadata}
\alias{visualize_metadata}
\title{Plot predefined figures of metadata}
\usage{
visualize_metadata(filepath.metadata.csv, dirname.target.figures)
}
\arguments{
\item{filepath.metadata.csv}{Character. Specify file path to csv containing GUANO metadata.}

\item{dirname.target.figures}{Character. Specify file path to to the target directory for the figures.}
}
\description{
Plot predefined figures of metadata
}
