% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_log.R
\name{write_log}
\alias{write_log}
\title{Write a log file of the specified parameters}
\usage{
write_log(params, filepath.logfile = NULL)
}
\arguments{
\item{params}{Object of class data.frame. A data frame containing information about specified parameters.}

\item{filepath.logfile}{Character. Specify file path to a log file. A new file will be written, if missing.}
}
\description{
Write a log file of the specified parameters
}
